Refactor main functions to return anyhow::Result for better error handling
- Updated the main function signatures across multiple examples to return `anyhow::Result<()>`.
- Added error handling for various run functions to propagate errors using the `?` operator.
- Improved command enum names for clarity in the tar example.
- Cleaned up unnecessary error handling patterns and ensured consistent error reporting.

+++ bk/crates/cats/algorithms/examples/graph/petgraph_astar.rs
@@ -27,7 +27,7 @@ fn main() {
         |node| {
             let (x, y) = graph[node];
             let (ex, ey) = graph[e];
-            (x.abs_diff(ex) + y.abs_diff(ey)) as u32
+            x.abs_diff(ex) + y.abs_diff(ey)
         },
     );
